Make the positions etc. in our buffer abstraction be longs, since we're using buffer's for everything and dictionaries/postings can pretty easily make it up over 2GB.  For now the in-memory buffers are limited to 2GB in size, but they will throw exceptions if asked to exceed their limits.  We need some big-big in-memory buffers, but that's for later.

+++ src/com/sun/labs/minion/indexer/dictionary/DiskDictionaryBundle.java
@@ -666,7 +666,7 @@ public class DiskDictionaryBundle<N extends Comparable> {
             throws java.io.IOException {
 
         DictionaryOutput fieldDictOut = mergeState.partOut.getPartitionDictionaryOutput();
-        int headerPos = fieldDictOut.position();
+        long headerPos = fieldDictOut.position();
         FieldHeader mergeHeader = new FieldHeader();
         mergeHeader.fieldID = mergeState.info.getID();
         mergeHeader.maxDocID = mergeState.maxDocID;
@@ -883,7 +883,7 @@ public class DiskDictionaryBundle<N extends Comparable> {
                 // We need a disk dictionary for calculating the lengths, so we'll
                 // open the one that we just wrote.  We'll start by remembering where 
                 // we were in the file!
-                int mdsp = fieldDictOut.position();
+                long mdsp = fieldDictOut.position();
 
                 File[] postOutFiles = mergeState.partOut.getPostingsFiles();
                 RandomAccessFile[] mPostRAF =
@@ -920,7 +920,7 @@ public class DiskDictionaryBundle<N extends Comparable> {
 
         //
         // Now zip back and write the header.
-        int endPos = fieldDictOut.position();
+        long endPos = fieldDictOut.position();
         fieldDictOut.position(headerPos);
         mergeHeader.write(fieldDictOut);
         fieldDictOut.position(endPos);
