+++ src/com/sun/labs/minion/util/buffer/NIOFileReadableBuffer.java
@@ -110,7 +110,7 @@ public class NIOFileReadableBuffer extends StdReadableImpl {
      */
     public NIOFileReadableBuffer(RandomAccessFile raf,
             long offset,
-            int limit) {
+            long limit) {
         this(raf, offset, limit, DEFAULT_BUFF_SIZE);
     }
 
@@ -127,7 +127,7 @@ public class NIOFileReadableBuffer extends StdReadableImpl {
      */
     public NIOFileReadableBuffer(RandomAccessFile raf,
             long offset,
-            int limit,
+            long limit,
             int buffSize) {
         this.raf = raf;
         chan = raf.getChannel();
@@ -140,7 +140,7 @@ public class NIOFileReadableBuffer extends StdReadableImpl {
         // Fill the buffer if the size of the data is smaller than the size
         // of the buffer.
         if(limit > 0 && limit <= buffSize) {
-            buff = ByteBuffer.allocate(limit);
+            buff = ByteBuffer.allocate((int) limit);
             int n = read(bs);
             ms = bs;
             me = bs + n;
@@ -190,8 +190,8 @@ public class NIOFileReadableBuffer extends StdReadableImpl {
      * Returns the number of bytes remaining to be read in the buffer.
      * @return The number of bytes remaining in the buffer.
      */
-    public int remaining() {
-        return (int) (be - pos);
+    public long remaining() {
+        return be - pos;
     }
 
     /**
@@ -218,7 +218,7 @@ public class NIOFileReadableBuffer extends StdReadableImpl {
      * buffer.  The first position in the sliced buffer is the given
      * position, and the limit on the sliced buffer is the given size.
      */
-    public ReadableBuffer slice(int p, int s) {
+    public ReadableBuffer slice(long p, long s) {
         return new NIOFileReadableBuffer(raf, bs + p, s, buff.capacity());
     }
 
@@ -226,15 +226,15 @@ public class NIOFileReadableBuffer extends StdReadableImpl {
      * Gets the limit of this buffer, i.e., the last readable position.
      * @return The last readable position in this buffer.
      */
-    public int limit() {
-        return (int) (be - bs);
+    public long limit() {
+        return be - bs;
     }
 
     /**
      * Sets the limit of this buffer, i.e., the last readable position.
      * @param l The limit that we wish to set for the buffer.
      */
-    public void limit(int l) {
+    public void limit(long l) {
         be = bs + l;
     }
 
@@ -243,7 +243,7 @@ public class NIOFileReadableBuffer extends StdReadableImpl {
      * @param i The position from which we wish to get a byte.
      * @return The byte at the given position.
      */
-    public byte get(int i) {
+    public byte get(long i) {
         return buff.get(checkBounds(i + bs));
     }
 
@@ -260,15 +260,15 @@ public class NIOFileReadableBuffer extends StdReadableImpl {
      * Gets the position of the buffer.
      * @return The current position in the buffer.
      */
-    public int position() {
-        return (int) (pos - bs);
+    public long position() {
+        return pos - bs;
     }
 
     /**
      * Positions the buffer.
      * @param i The position to which we should set the buffer.
      */
-    public void position(int i) {
+    public void position(long i) {
         this.pos = bs + i;
     }
